docs: typo in install commands (win & termux) - README.md (#1051)
In Install From Source, for both Windows and Termux, 
the last command 'rm -rf /ani-cli' doesn't work,
since there is no 'ani-cli' folder in the '/' directory. It
doesn't throw error because of the -f flag, but it also
fails to remove the intended folder. This commit 
fixes this by removing the slash. Previous commands 
use 'ani-cli' instead of './ani-cli', so removing the slash 
seems to make more sense than adding a dot before it.